pyproject.toml
[project]
name = "S16"
version = "0.3.0"
description = "Agentic AI - Multi-Agent Orchestration System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [

    "pyyaml",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.3",
    "asyncio>=3.4.3",

    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "httpx>=0.28.1",
    "llama-index>=0.12.31",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pymupdf4llm>=0.0.21",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "trafilatura[all]>=2.0.0",
    "jinja2>=3.1.6",
    "playwright>=1.52.0",
    "readability-lxml>=0.8.4.1",

    "networkx>=3.4.2",
    "json-repair>=0.44.1",

    "markdownify>=1.1.0",
    "uvicorn>=0.34.2",
    "fastapi>=0.115.12",
    "sse-starlette>=2.3.4",
    "feedparser>=6.0.12",
    "beautifulsoup4>=4.13.4",
    "aiohttp>=3.11.18",
    "watchdog>=3.0.0",
    "langchain-ollama>=0.3.2",
    "langchain-google-genai>=2.1.2",
    "rank_bm25>=0.2.2",
]
